requirements.txt
streamlit==1.29.0
pillow==10.2.0
requests==2.31.0
matplotlib==3.8.2
google-cloud-firestore==2.14.0